+++ dash/dash.py
@@ -789,15 +789,14 @@ class Dash:
             }
         )
 
-    def serve_dist(self):
-        libraries = flask.request.get_json()
+    def get_dist(self, libraries):
         dists = []
         for dist_type in ("_js_dist", "_css_dist"):
             resources = ComponentRegistry.get_resources(dist_type, libraries)
             srcs = self._collect_and_register_resources(resources, False)
             for src in srcs:
                 dists.append(dict(type=dist_type, url=src))
-        return flask.jsonify(dists)
+        return dists
 
     def _collect_and_register_resources(self, resources, include_async=True):
         # now needs the app context.
@@ -1254,8 +1253,6 @@ class Dash:
     def dispatch(self):
         body = flask.request.get_json()
 
-        nlibs = len(ComponentRegistry.registry)
-
         g = AttributeDict({})
 
         g.inputs_list = inputs = body.get(  # pylint: disable=assigning-non-slot
@@ -1285,7 +1282,6 @@ class Dash:
 
         try:
             cb = self.callback_map[output]
-            _allow_dynamic = cb.get("allow_dynamic_callbacks", False)
             func = cb["callback"]
             g.background_callback_manager = (
                 cb.get("manager") or self._background_manager
@@ -1361,16 +1357,11 @@ class Dash:
                     outputs_list=outputs_list,
                     long_callback_manager=self._background_manager,
                     callback_context=g,
+                    app=self,
                     app_on_error=self._on_error,
                 )
             )
         )
-
-        if not _allow_dynamic and nlibs != len(ComponentRegistry.registry):
-            print(
-                "Warning: component library imported during callback, move to top-level for full support.",
-                file=sys.stderr,
-            )
         return response
 
     def _setup_server(self):